pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "afs-cortex"
version = "0.1.0"
description = "AFS parser — Snowflake Cortex edition"
requires-python = ">=3.10"
dependencies = [
    "snowflake-ml-python>=1.5.0",
    "snowflake-snowpark-python",
    "pydantic>=2.0",
    "rapidfuzz",
]

[tool.setuptools.packages.find]
where = ["src"]
